+++ CHANGELOG.rst
@@ -13,7 +13,7 @@ Release 0.14.0 (unreleased)
 * Add new ``remove`` command to remove projects from manifest and disk (#26)
 * Fix "unsafe symlink target" error for archives containing relative ``..`` symlinks (#1122)
 * Fix ``dfetch add`` crashing with a ``ValueError`` when the remote URL has a trailing slash (#1137)
-* Fix unhelpful message when a metadata file (#1145)
+* Fix unhelpful error message when a metadata file is malformed (#1145)
 
 Release 0.13.0 (released 2026-03-30)
 ====================================